django/requirements.txt
Django==1.11.4
django-admin==1.0.4
django-filter==1.0.4
django-views==1.0.0
djangorestframework==3.6.3
ExifRead==2.1.2
Markdown==2.6.9
pytz==2017.2
simplekml==1.3.0
